requirements.txt
openai>=1.0.0
pyyaml>=6.0
networkx>=3.0
matplotlib>=3.5.0
